requirements.txt
celery>=5.2.0,<5.2.7
Django>=3.2.16, <4.1.2
djangorestframework-simplejwt>=4.0.0, <5.2.1
djangorestframework>=3.12.4,<3.14.0
dj_rest_auth>=2.1.3,<2.2.5
django-environ>=0.7.0,<0.9.0
django-allauth>=0.40.0,<0.51.0
django-cors-headers>=3.2.0, <3.13.0
django-rest-auth>=0.9.0 ,<0.9.5
django-kronos>=0.90, <1.10
django-dbbackup==4.0.2
django-redis>=4.0.0, <5.2.0
django-sendgrid-v5>=1.1.0,<1.2.1
django-rest-passwordreset>=1.1.0, <1.3.0
django-axes>=5.11.0, <5.39.0
django-debug-toolbar>=3.2, <3.7.0
django-imagekit>=4.0,<4.1.0
django-cleanup>=5.0.0,<6.0.0
django-phonenumber-field>=5.0.0, <7.0.0
gunicorn>=20.0.0, <20.1.0
google-auth>=2.4.1, <2.12.0
google-api-python-client>=2.33.0,<2.64.0
google-analytics-data>=0.11.0<0.14.1
jaconv>=0.2<0.3
flake8>=3.9.0, <5.0.4
PyYAML>=5.4.1, <6.0
phonenumbers>=8.12.35, <8.12.56
pillow>=8.2.0,<9.2.0
psycopg2>=2.8.6, <2.9.3
psycopg2-binary>2.8, <2.9.3
python-dotenv>=0.19, <0.21.0
stripe>=2.48.0,<4.2.0
python-dateutil>=2.7.5,<2.8.2
redis>=4.1.0, <4.3.4
social-auth-app-django>=2.0.0, <5.0.0
sendgrid>=6.9.0,<6.9.7
sentry-sdk>=1.5.0,<1.9.10
oauth2client>=4.0.0,<4.1.3
ndg-httpsclient>=0.4.4<0.5.1
pyOpenSSL>=21.0.0<22.1.0
boto3>=1.24.90<1.24.94
django-storages>=1.12.1<1.13.1